Moved ensure_metafits into metadb_utils
